requirements.txt
aiohttp
kurigram
python-dotenv
tgcrypto
motor
aiofiles
dnspython
huggingface_hub
